Extract layout module without Material dependency, move Spacer there
Spacer.kt only used compose.foundation APIs, so it didn't need to live
in uikit (which depends on material3).
